Move function and cleanup
